requirements/prd.txt
PyGithub==2.7.0
PyYAML>=5.4.0
requests==2.32.5
pytest==8.4.1
Flask==1.1.4
itsdangerous==1.1.0
Werkzeug==0.16.1
